backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
astropy==7.0.0
astropy-iers-data==0.2024.12.9.0.36.21
asyncpg==0.30.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.6
fonttools==4.55.3
greenlet==3.1.1
h11==0.14.0
healpy==1.18.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pyerfa==2.0.1.5
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.32.1
